Fixed some servers stuff.
